Tests improvements (#636)
* increase line length to 100

* generalize tests

* test aspect

* added conftest.py

* assert numpy equal dask cupy case

* test classify

* curvature to return float32

* test curvature

* test dtype of output

* classify to return float32

* set verify_dtype to False as default

* test focal

* test hillshade

* aspect, classify, curvature: verify dtype of output array

* test multispectral

* linting

* added rtol as a param

* test pathfinding

* allocation: return dtype of float32

* test proximity

* flake8

* test slope

* test terrain

* test viewshed

* test zonal

* change as requested

* Fix cupy and rtxpy tests

+++ xrspatial/proximity.py
@@ -885,7 +885,7 @@ def allocation(
 
     # convert to have same type as of input @raster
     result = xr.DataArray(
-        (allocation_img).astype(raster.dtype),
+        allocation_img,
         coords=raster.coords,
         dims=raster.dims,
         attrs=raster.attrs,
